shutdown: Move all umountable mounts before unmounting them
There is a case that confuses systemd-shutdown: a filesystem has been moved to
a mount point which is part of another filesystem from an image from that
former filesystem. systemd-shutdown cannot unmount any of those two
filesystems. It needs first to move the filesystem containing the image of the
other out of the tree of that image.

Here we move all unmountable mount points to sub-directories within the same
directory `/run/shutdown/mounts`. That way no cycle can exist.